unify model opts; add attention rescore in decodable; rename ds2 ctc beam search
